Move logic hooks into /hooks folder
